[FEAT] 화면 크기 조절에 따른 상품 갯수 변화 함수 custom hook으로 변경
